Update and rename SettingsConnectionsController.kt to SettingsConnectionsScreen.kt
